Rename .csproj and namespaces for 3.x sample
