+++ aspnetcore/web-api/advanced/custom-formatters/samples/3.x/CustomFormattersSample/Controllers/ContactsController.cs
@@ -1,15 +1,12 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using CustomFormatterDemo.Models;
 using System.Collections.Concurrent;
+using System.Collections.Generic;
+using CustomFormattersSample.Models;
+using Microsoft.AspNetCore.Mvc;
 
-
-namespace CustomFormatterDemo.Controllers
+namespace CustomFormattersSample.Controllers
 {
-   // [ApiController]
+    // [ApiController]
     [Route("api/[controller]")]
     public class ContactsController : Controller
     {
