+++ aspnetcore/web-api/advanced/custom-formatters/samples/3.x/CustomFormattersSample/Models/Contact.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace CustomFormatterDemo.Models
+﻿namespace CustomFormattersSample.Models
 {
     public class Contact
     {
