+++ aspnetcore/web-api/advanced/custom-formatters/samples/3.x/CustomFormattersSample/Startup.cs
@@ -1,12 +1,11 @@
-﻿using CustomFormatterDemo.Formatters;
-using CustomFormatterDemo.Models;
+﻿using CustomFormattersSample.Formatters;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace CustomFormatterDemo
+namespace CustomFormattersSample
 {
     public class Startup
     {
